Replace float64 with Float type in mat package
wherever possible

+++ pkg/mat/dense_formatting.go
@@ -145,7 +145,7 @@ func (d *Dense) formattingMaxColumnsWidth(
 	return widths, maxWidth
 }
 
-func formatValue(buf []byte, val float64, c rune, precision int) []byte {
+func formatValue(buf []byte, val Float, c rune, precision int) []byte {
 	return strconv.AppendFloat(buf[:0], val, byte(c), precision, 64)
 }
